Add AAPT2 support for robolectric tests
