+++ .travis.yml
@@ -18,3 +18,6 @@ android:
     - android-26
 
 script: ./gradlew build
+
+after_failure:
+  - cat app/build/reports/tests/testDebugUnitTest/index.html
