+++ easypermissions/src/test/java/pub/devrel/easypermissions/EasyPermissionsTest.java
@@ -1,18 +1,14 @@
-package pub.devrel.easypermissions.sample;
+package pub.devrel.easypermissions;
 
 import android.Manifest;
-import android.content.Context;
+import android.app.Application;
 
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
-import org.robolectric.Shadows;
+import org.robolectric.RuntimeEnvironment;
 import org.robolectric.annotation.Config;
 import org.robolectric.shadows.ShadowApplication;
-
-import pub.devrel.easypermissions.EasyPermissions;
 
 import static junit.framework.Assert.assertFalse;
 import static junit.framework.Assert.assertTrue;
@@ -23,19 +19,9 @@ import static junit.framework.Assert.assertTrue;
 @RunWith(RobolectricTestRunner.class)
 @Config(constants = BuildConfig.class, sdk = 23)
 public class EasyPermissionsTest {
-
-    private MainActivity mActivity;
-    private ShadowApplication mApplication;
-
-    @Before
-    public void setup() {
-        mActivity = Robolectric.buildActivity(MainActivity.class).get();
-        mApplication = Shadows.shadowOf(mActivity.getApplication());
-    }
-
     @Test
     public void testHasPermissions() {
-        Context context = mApplication.getApplicationContext();
+        Application app = RuntimeEnvironment.application;
 
         String[] perms = new String[]{
                 Manifest.permission.READ_SMS,
@@ -43,14 +29,14 @@ public class EasyPermissionsTest {
         };
 
         // Wes should not have permissions before any are granted
-        assertFalse(EasyPermissions.hasPermissions(context, perms));
+        assertFalse(EasyPermissions.hasPermissions(app, perms));
 
         // Granting one permission should not make the whole set appear granted
-        mApplication.grantPermissions(perms[0]);
-        assertFalse(EasyPermissions.hasPermissions(context, perms));
+        ShadowApplication.getInstance().grantPermissions(perms[0]);
+        assertFalse(EasyPermissions.hasPermissions(app, perms));
 
         // Granting all permissions should make the whole set granted
-        mApplication.grantPermissions(perms);
-        assertTrue(EasyPermissions.hasPermissions(context, perms));
+        ShadowApplication.getInstance().grantPermissions(perms);
+        assertTrue(EasyPermissions.hasPermissions(app, perms));
     }
 }
