+++ gradle.properties
@@ -17,6 +17,3 @@ org.gradle.jvmargs=-Xmx1536m
 # This option should only be used with decoupled projects. More details, visit
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
-
-# https://github.com/robolectric/robolectric/issues/3157
-android.enableAapt2=false
